+++ src/api_insee/conf.py
@@ -9,6 +9,7 @@ ApiPathName = Literal[
     "path_token",
     "path_siren",
     "path_siret",
+    "path_informations",
     "path_liens_succession",
 ]
 
@@ -17,6 +18,7 @@ class ApiUrls(TypedDict):
     path_token: str
     path_siren: str
     path_siret: str
+    path_informations: str
     path_liens_succession: str
 
 
@@ -30,5 +32,6 @@ class ApiVersion(Enum):
             "path_token": _TOKEN_URL,
             "path_siren": f"{self.value}/siren",
             "path_siret": f"{self.value}/siret",
+            "path_informations": f"{self.value}/informations",
             "path_liens_succession": f"{self.value}/siret/liensSuccession",
         }