+++ src/api_insee/request/request_informations.py
@@ -1,0 +1,5 @@
+from api_insee.request.request import RequestService
+
+
+class RequestInformationsService(RequestService):
+    path_name = "path_informations"